Add files via upload
int, char, Boolean, and float are primitives. String is non-primitive. What about students variable? This is an object of the user-created class Student. It was created simply to illustrate the work with non-primitive variables. They are the majority in Java, since almost everything in this programming language is an object. For understanding, here is the code of the Student class: